Add configuration by env variables

+++ internal/middlewares/rabbitmq/fanout.go
@@ -9,7 +9,7 @@ import (
 
 type FanoutPublisherConfig struct {
 	Exchange string
-	Timeout  uint
+	Timeout  uint8
 }
 
 type FanoutPublisher struct {
